+++ CHANGELOG.md
@@ -268,3 +268,4 @@
 * Rename `Flat` morphism to `Straight`
 * Rename `Dual` morphism to `Opposite`
 * Define `I` datatype for type level natural numbers
+* Define `Nonempty` type family